+++ Test/RivetTest/Functions/New-Migration.ps1
@@ -1,13 +1,13 @@
 
 function New-TestMigration
 {
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName='NamedMigration')]
     param(
         [Parameter(Mandatory,ValueFromPipeline,Position=0)]
         # The migration contents to output.
         [String]$InputObject,
 
-        [Parameter(Mandatory)]
+        [Parameter(Mandatory, ParameterSetName='NamedMigration')]
         [Alias('Name')]
         # The name of the migration.
         [String]$Named,
@@ -15,7 +15,10 @@ function New-TestMigration
         # The name of the database.
         [String]$DatabaseName = $RTDatabaseName,
 
-        [String]$ConfigFilePath = $RTConfigFilePath
+        [String]$ConfigFilePath = $RTConfigFilePath,
+
+        [Parameter(Mandatory, ParameterSetName='SchemaPs1')]
+        [switch] $AsCheckpoint
     )
 
     Set-StrictMode -Version 'Latest'
@@ -27,18 +30,35 @@ function New-TestMigration
         New-Item -Path $migrationsRoot -ItemType 'Directory' -Force | Format-Table | Out-String | Write-Verbose
     }
 
-    do
+    if ($AsCheckpoint)
     {
-        $script:RTTimestamp++
-        $migrationFileName = "${script:RTTimestamp}_*.ps1"
-        $migrationPath = Join-Path -Path $migrationsRoot -ChildPath $migrationFileName
+        $migrationFileName = 'schema.ps1'
     }
-    while( (Test-Path -Path $migrationPath -PathType Leaf) )
+    else
+    {
+        do
+        {
+            $script:RTTimestamp++
+            $migrationFileName = "${script:RTTimestamp}_*.ps1"
+            $migrationPath = Join-Path -Path $migrationsRoot -ChildPath $migrationFileName
+        }
+        while( (Test-Path -Path $migrationPath -PathType Leaf) )
 
-    $migrationFileName = "${script:RTTimestamp}_${Named}.ps1"
+        $migrationFileName = "${script:RTTimestamp}_${Named}.ps1"
+    }
+
     $migrationPath = Join-Path -Path $migrationsRoot -ChildPath $migrationFileName
     $InputObject | Set-Content -Path $migrationPath
-    Get-Item -Path $migrationPath
+    $migrationFile = Get-Item -Path $migrationPath
+
+    if (-not $AsCheckpoint)
+    {
+        $migrationFile |
+            Add-Member -Name 'MigrationID' -MemberType NoteProperty -Value $script:RTTimestamp -PassThru |
+            Add-Member -Name 'MigrationName' -MemberType NoteProperty -Value $Named
+    }
+
+    return $migrationFile
 }
 
 Set-Alias -Name 'GivenMigration' -Value 'New-TestMigration'